% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enkpf.R
\name{adaptive_gamma}
\alias{adaptive_gamma}
\title{Select gamma adaptively}
\usage{
adaptive_gamma(e.0 = 0.5, e.1 = 0.8, imax = 15, imin = 1,
  delta.gam = 1/(imax - imin + 1), y = y, ...)
}
\arguments{
\item{e.0}{ESS lower bound}

\item{e.1}{ESS upper bound}

\item{imax, }{imin: gammas considered go from imin/(imax-imin+1) to imax/(imax-imin+1),
by default from 1/15 to 1}

\item{delta.gam}{gammas increments, can be changed}

\item{y, }{can be used potentially, but not currently.}

\item{...}{additional parameters passed to enkf_step}
}
\value{
best_model list containing computed quantities for the optimal gamma chosen
}
\description{
Binary search such that e.0 <= ess <= e.1
}
