% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enkpf.R
\name{bal.sample}
\alias{bal.sample}
\title{Balanced sampling}
\usage{
bal.sample(w, R = length(w), unif = runif(1))
}
\arguments{
\item{w}{= probabilities (must sum to one !)}

\item{R}{= sample size}

\item{unif}{= uniform (pass if deterministic behaviour wished)}
}
\value{
list(N, index)
}
\description{
Multiplicities N_1, ..., N_n, i.e. element j is sampled N_j times. 
index = numbers of sampled elements

Author: Hans-Rudolf Kuensch, Date: 21 Apr 2010.
Modified by Sylvain Robert (added unif as argument and some names)
}
