% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_block_LEnKPF.R
\name{block_LEnKPF}
\alias{block_LEnKPF}
\title{block-LEnKPF analysis}
\usage{
block_LEnKPF(xb, y, H, R, l, block_size = l/2,
  get_partition = ring_partition, ndim = nrow(xb), taper = 1,
  gam.fix = NA, d = length(y), K = ncol(xb), q = nrow(xb),
  unif = runif(1), ...)
}
\arguments{
\item{xb}{the background ensemble}

\item{y}{the observations}

\item{H}{the observation linear operator}

\item{R}{the observations error covariance}

\item{l}{the localization radius (not used directly)}

\item{block_size}{size of domain considered to include observations in a block}

\item{get_partition}{function to partition the observations in blocks, depending on geometry (typically ring_partition or sweq_partition)}

\item{taper}{the tapering correlation matrix applied to P}

\item{gam.fix}{fixed gamma value}

\item{d}{dimension of observation, ensemble and state space}

\item{K}{dimension of observation, ensemble and state space}

\item{q}{dimension of observation, ensemble and state space}

\item{unif}{used for balanced sampling}

\item{...}{additional parameters passed to adaptive_gamma, typically e.0 and e.1}
}
\description{
Assimilate observations sequentially in blocks. 
Smooth discontinuities by conditional resampling using the empirical covariance structure.
The method does not work without a taper for P. 
The specific update of one block is in one_block_update
REM: for historical reasons the indices  referred to as u and v in the paper are named in the code as x and v.
}
