% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cycle.R
\name{da_cycle}
\alias{da_cycle}
\title{Function for cycled data assimilation}
\usage{
da_cycle(xb, model.run, f.update, rho = 1, nonoise = FALSE,
  verbose = TRUE, manifold_proj = NULL, cov_inflation = NULL, ...)
}
\arguments{
\item{xb}{the initial ensemble}

\item{model.run}{the true model time series}

\item{f.update}{the analysis function}

\item{rho}{the covariance inflation parameter}

\item{nonoise}{if TRUE doesn't inject noise in forecast step (bad idea)}

\item{verbose}{if TRUE minimal print info}

\item{manifold_proj}{is to project the ensemble on a manifold after analysis, typicall sweq_proj}

\item{cov_inflation}{for creating customized inflation function. The default is just multiplicative inflation}

\item{...}{extra argument for f.update}
}
\value{
a list with the xa time series, and other informations returned by f.update
}
\description{
Function for cycled data assimilation
}
