% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cycle.R
\name{da_update}
\alias{da_update}
\title{Wrapper function for assimilation algorithms}
\usage{
da_update(method = c("EnKPF", "EnKF", "PF", "LEnKF", "block-LEnKF",
  "naive-LEnKPF", "block-LEnKPF", "LPF", "FF"),
  get_neighbours = ring_neighbours, get_partition = ring_partition,
  taper = 1, l = 10, ndim = ndim, e.0 = 0.5, e.1 = 0.8, ...)
}
\arguments{
\item{method}{is the name of the algorithm to use}

\item{get_neighbours}{the function to get neighboring points, depending on geometry (typically ring_neighbours or sweq_neighbours)}

\item{get_partition}{function to partition the observations in blocks, depending on geometry (typically ring_partition or sweq_partition)}

\item{taper}{the tapering correlation matrix applied to P}

\item{l}{the localization radius (not used directly)}

\item{...}{additional parameters passed to adaptive_gamma, typically e.0 and e.1}
}
\description{
Wrapper function to call the individual algorithms by a name.
Useful for simulations.
}
