% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enkpf.R
\name{enkf_step}
\alias{enkf_step}
\title{EnKF part of EnKPF}
\usage{
enkf_step(y, xb, gam, P, H, R, y.resx, kloc = FALSE, ...)
}
\arguments{
\item{y}{the observations}

\item{xb}{the background ensemble}

\item{gam}{gamma parameter}

\item{P}{background covariance matrix}

\item{H}{observation operator}

\item{R}{observation covariance}

\item{y.resx}{= y - Hxb the innovations}

\item{kloc}{if TRUE use the localized Kalman gain formulation}

\item{...}{additional arguments used in case kloc=TRUE (see kloc_gain)}
}
\value{
a list with Kalman gains, weights and other quantities of interest
}
\description{
Compute the non-random elements of the EnKPF update
typically used for choosing gamma adaptively
}
