% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_EnKPF.R
\name{EnKPF}
\alias{EnKPF}
\title{Global EnKPF}
\usage{
EnKPF(xb, y, H, R, taper = 1, P = cov(t(xb)) * taper, gam.fix = NA,
  d = length(y), K = ncol(xb), q = nrow(xb), R2 = msqrt(R), eps1 = R2
  \%*\% matrix(rnorm(d * K), d, K), eps2 = R2 \%*\% matrix(rnorm(d * K), d,
  K), unif = runif(1), ...)
}
\arguments{
\item{xb}{the background ensemble}

\item{y}{the observations}

\item{H}{the observation linear operator}

\item{R}{the observations error covariance}

\item{taper}{the tapering correlation matrix applied to P}

\item{P}{the background covariance}

\item{gam.fix}{fixed gamma value}

\item{d, K, q}{dimension of observation, ensemble and state space}

\item{R2}{matrix square-root of R}

\item{eps1, eps2}{random perturbations N(0,R)}

\item{unif}{used for balanced sampling}

\item{...}{additional parameters passed to adaptive_gamma, typically e.0 and e.1}
}
\value{
a list with xa, gam and ess
}
\description{
Special cases: 
setting gam.fix=1: EnKF
setting gam.fix=0: PF
setting gam.fix=1 and kloc=TRUE recovers the LEnKF

Otherwise gamma is chosen adaptively (then pass e.0 and e.1 as arguments)
see also utils_enkpf for the exact procedure.
}
