% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enkpf.R
\name{ensure_PD}
\alias{ensure_PD}
\title{To ensure positive definiteness}
\usage{
ensure_PD(P)
}
\description{
Only used in pathological cases.
Project covariance P on closed positive-definite matrix.
Used in enkf_step to ensure that Q is pd, which happens to not be the case
in a few pathological cases.
}
