% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{f_crps}
\alias{f_crps}
\title{Compute CRPS score}
\usage{
f_crps(smple, obs)
}
\arguments{
\item{smple}{predictive ensemble (a vector)}

\item{obs}{observed value (a scalar)}
}
\value{
the crps score
}
\description{
CRPS score for a predictive ensemble (O(n*log(n)) Variante)
(Author: Marco)
}
\examples{
set.seed(1)
truth <- 0
n <- 100
f_crps( rnorm(n, truth, 1), truth)
f_crps( rnorm(n, truth, 0.1), truth)
f_crps( rnorm(n, truth+2, 1), truth)
ens.means <- rnorm(n, truth, 1)
plot(density(ens.means)); abline(v=truth)
}
