% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{f_crps_wgm}
\alias{f_crps_wgm}
\title{Compute CRPS score for a Gaussian mixture}
\usage{
f_crps_wgm(mu, sig2, obs, w)
}
\arguments{
\item{mu}{mixture means}

\item{sig2}{mixture variance}

\item{obs}{observations}

\item{w}{mixture component probability}
}
\value{
the crps score
}
\description{
unequal weights case
}
\examples{
set.seed(1)
truth <- 0
n <- 20
ens.means <- rnorm(n, truth, 1)
w <- dnorm( ens.means - truth)
w <- w/sum(w)
f_crps_wgm(mu=ens.means, sig2=.1, obs=truth, w=w)
sigvec <- rgamma(n, 1)
f_crps_wgm(mu=ens.means, sig2=sigvec, obs=truth, w=w)
f_crps_wgm(  mu=ens.means, sig2=.1, obs=truth, w=rep(1/n, n))
}
