% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{GC_function}
\alias{GC_function}
\title{GC covariance function}
\usage{
GC_function(z, c)
}
\arguments{
\item{z}{is a distance}

\item{c}{the support half-length}
}
\description{
correlation function from Gaspari & Coh 1999, eq. 4.10
used in GC_taper
}
\examples{
GC_function(2, 5)
sapply(1:10, GC_function, c=5)
}
