% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics_gp.R
\name{gp_simulate}
\alias{gp_simulate}
\title{Simulate a correlated Gaussian random field}
\usage{
gp_simulate(q = 100, l = 5, Rsig2 = 1)
}
\arguments{
\item{q}{dimension}

\item{l}{half correlation length for the covariance matrix}

\item{Rsig2}{variance of the observation}
}
\value{
list with x the field, y the observations and various informations about the processes used for assimilation
}
\description{
Simulate a correlated Gaussian random field. The covariance function is the GC (Gaspari-Cohn) with half length determined by l
Observations are made at every location with independent errors and variance Rsig2
}
\examples{
gp_data <- gp_simulate(q=100)
xb <- gp_ens0(20, gp_data)
gp_plot(gp_data, xb,  'Background')
}
