% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enkpf.R
\name{kalman.gain}
\alias{kalman.gain}
\title{Compute the Kalman Gain}
\usage{
kalman.gain(P, H, R)
}
\arguments{
\item{P}{the covariance matrix}

\item{H}{the observation linear operator}

\item{R}{the observations error covariance}
}
\value{
K
}
\description{
Uses the cholesky decomposition of (HPH' + R)
}
