% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enkpf.R
\name{kloc_gain}
\alias{kloc_gain}
\title{Equivalent local Kalman gain}
\usage{
kloc_gain(P, H, R, l, get_neighbours, ndim = nrow(P))
}
\arguments{
\item{P}{the covariance matrix}

\item{H}{the observation linear operator}

\item{R}{the observations error covariance}

\item{l}{the localization radius}

\item{get_neighbours}{the function to get neighboring points
(typically ring_neighbours or sweq_neighbours)}

\item{ndim}{different from q in case of sweq}
}
\description{
compute a Kalman gain locally at each location and construct a global
Kalman gain with the appropriate values and zeroes
}
