% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enkpf.R
\name{leftmatch}
\alias{leftmatch}
\title{Perumute indices}
\usage{
leftmatch(leftind, rightind)
}
\arguments{
\item{leftind:}{indices on the left (reference)}

\item{rightind:}{indices on the right}
}
\value{
the right indices rearranged
}
\description{
Used to match particles between local updates.
Rearrange indices on the right to match as many as
possible on the left... Default for left should be 1:length(rightind)
}
\examples{
leftind <- c(1,2,3,3,3)
rightind<- c(1,3,3,4,5)

leftmatch(leftind, rightind)
}
