% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enkpf.R
\name{msqrt}
\alias{msqrt}
\title{Symmetric matrix square root of a positive definite matrix}
\usage{
msqrt(A)
}
\arguments{
\item{A}{a symmetric matrix}
}
\value{
S such that S'S = A
}
\description{
use eigenvalue decompostion
}
