% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{multiplot}
\alias{multiplot}
\title{Combine multiple ggplots}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{can be any number of plots as returned by ggplot}

\item{plotlist}{is alternatively a list of plots}

\item{cols}{is the number of columns in which to organize the plots}

\item{layout}{can alternatively be set manually}
}
\description{
convenient function to combine multiple ggplot plots on one page
}
