% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_naive_LEnKPF.R
\name{naive_LEnKPF}
\alias{naive_LEnKPF}
\title{naive LEnKPF analysis}
\usage{
naive_LEnKPF(xb, y, H, R, l, get_neighbours = ring_neighbours,
  ndim = nrow(xb), taper = 1, gam.fix = NA, d = length(y),
  K = ncol(xb), q = nrow(xb), unif = runif(1), R2 = msqrt(R),
  eps1 = R2 \%*\% matrix(rnorm(d * K), d, K), eps2 = R2 \%*\% matrix(rnorm(d
  * K), d, K), ...)
}
\arguments{
\item{xb}{the background ensemble}

\item{y}{the observations}

\item{H}{the observation linear operator}

\item{R}{the observations error covariance}

\item{l}{the localization radius}

\item{get_neighbours}{the function to get neighboring points, depending on geometry (typically ring_neighbours or sweq_neighbours)}

\item{taper}{the tapering correlation matrix applied to P}

\item{gam.fix}{fixed gamma value}

\item{d}{dimension of observation, ensemble and state space}

\item{K}{dimension of observation, ensemble and state space}

\item{q}{dimension of observation, ensemble and state space}

\item{unif}{used for balanced sampling}

\item{R2}{matrix square-root of R}

\item{eps1}{random perturbations N(0,R)}

\item{eps2}{random perturbations N(0,R)}

\item{...}{additional parameters passed to adaptive_gamma, typically e.0 and e.1}
}
\description{
Apply an independent EnKPF analysis at each location using only neighboring information.

Special cases:
setting gam.fix=1: LEnKF
setting gam.fix=0: LPF
}
