% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_block_LEnKPF.R
\name{one_block_update}
\alias{one_block_update}
\title{assimilation of one block with block-LEnKPF}
\usage{
one_block_update(xa, y, P, H, R, ind, K = ncol(xa), gam.fix = NA,
  unif = runif(1), ...)
}
\arguments{
\item{y}{the observations}

\item{P}{the background covariance}

\item{H}{the observation linear operator}

\item{R}{the observations error covariance}

\item{ind}{are the indices of used for the current block, containing x.ind, v.ind and y.ind}

\item{K}{dimension of observation, ensemble and state space}

\item{gam.fix}{fixed gamma value}

\item{unif}{used for balanced sampling}

\item{...}{additional parameters passed to adaptive_gamma, typically e.0 and e.1}
}
\description{
Assimilation of one block of observations. 
Gamma is chosen adaptively if not specified through gam.fix with adaptive gamma.
}
