% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{pit_hist}
\alias{pit_hist}
\title{Compute PIT Histogram}
\usage{
pit_hist(hx, y, fig = TRUE, ...)
}
\arguments{
\item{hx}{ensemble prediction (dxK :one row for each y, one col for each ensemble member)}

\item{y}{actual vector of observations}

\item{fig}{should the histogram be plotted or the ranks be returned?}
}
\value{
the vector of ranks (length of y)
}
\description{
PIT= Probability integral transform, also called Rank Histogram
}
\examples{
set.seed(1)
d <- 1000
yo <- rnorm(d)
k <- 40
pit <- pit_hist(matrix(rnorm(d*k, 0, 1  ), d, k), yo, main='calibrated ensemble')
pit <- pit_hist(matrix(rnorm(d*k, 0, 0.5), d, k), yo, main='underdispersed')
pit <- pit_hist(matrix(rnorm(d*k, 0, 1.5), d, k), yo, main='overdispersed')
}
