% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{pit_sweq_plot}
\alias{pit_sweq_plot}
\title{Plot PIT histogram}
\usage{
pit_sweq_plot(pit_df, K = 50)
}
\arguments{
\item{pit_df}{is a data frame produced by pit_sweq}
}
\description{
PIT histograms for sweq by field and type (background or analysis)
}
\examples{
set.seed(1)
sweq_run <- sweq_simulate(5*60, 60, 300)
k <- 20
ens0 <- sweq_ens0(k, sweq_run, klag=1000)
l <- 5
block_run <- da_cycle(ens0, sweq_run, block_LEnKPF, l=l, block_size=l/2, 
                      taper=sweq_GC(sweq_run$ndim, l, 0.9), ndim=sweq_run$ndim,
                      get_partition = sweq_partition,
                      manifold_proj = sweq_proj)
pit_df <- pit_sweq(sweq_run, block_run, fig=TRUE, train=train)
pit_df \%>\% pit_sweq_plot()                  
}
