% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_crps}
\alias{plot_crps}
\title{Plot time evolution of CRPS for all fields}
\usage{
plot_crps(ens_eval)
}
\arguments{
\item{ens_eval}{is a data frame produced by sweq_eval}
}
\description{
Plot time evolution of CRPS for all fields
}
\examples{
set.seed(1)
sweq_run <- sweq_simulate(5*60, 60, 300)
k <- 20
ens0 <- sweq_ens0(k, sweq_run, klag=1000)
l <- 5
block_run <- da_cycle(ens0, sweq_run, block_LEnKPF, l=l, block_size=l/2, 
                      taper=sweq_GC(sweq_run$ndim, l, 0.9), ndim=sweq_run$ndim,
                      get_partition = sweq_partition,
                      manifold_proj = sweq_proj)
block_df <- sweq_eval(sweq_run, block_run)
block_df \%>\% mutate(method='block-LEnKPF') \%>\% plot_crps()
}
