% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{ring_alldist}
\alias{ring_alldist}
\title{Pairwise distances on a rings}
\usage{
ring_alldist(N)
}
\arguments{
\item{N}{the size of the ring}
}
\value{
the matrix of pairwise distances on a ring of size N.
}
\description{
Pairwise distances on a rings
}
\examples{
A <- ring_alldist(9)
}
