% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{ring_dist}
\alias{ring_dist}
\title{Compute minimal distances (discrete) on a ring}
\usage{
ring_dist(i, j, N)
}
\arguments{
\item{i}{from}

\item{j}{to}

\item{N}{size of the ring}
}
\value{
distance on a ring between i and j
}
\description{
Compute minimal distances (discrete) on a ring
}
\examples{
ring_dist(1, 9, 10)
}
