% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{ring_GC}
\alias{ring_GC}
\title{GC taper on ring}
\usage{
ring_GC(ndim, c)
}
\arguments{
\item{ndim}{number of dimensions in physical space}

\item{c}{the support half-length}
}
\value{
taper matrix
}
\description{
Compute the GC taper on a ring, with support half-length c
}
\examples{
taper <- ring_GC(100, 10)
image_mat(taper)
}
