% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{ring_mod}
\alias{ring_mod}
\title{Ring modulo}
\usage{
ring_mod(a, b)
}
\arguments{
\item{a, b}{a mod b}
}
\description{
Special modulo for indices on ring ( a mod b)
with special treatment for a==0
}
\examples{
ring_mod(-2, 10)
}
