% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{ring_neighbours}
\alias{ring_neighbours}
\title{Neighborhood on a ring
get the l neighbours on each side of
i on a ring of size ndim}
\usage{
ring_neighbours(i, l, ndim)
}
\arguments{
\item{i}{center}

\item{l}{number of neighbours in each direction}

\item{ndim}{size of the ring}
}
\value{
list(x.ind, center, local.center)
local.center are the indices of the centers in the local space (x.local) to apply the update
}
\description{
Neighborhood on a ring
get the l neighbours on each side of
i on a ring of size ndim
}
