% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{ring_partition}
\alias{ring_partition}
\title{Partition the observations for the block-LEnKPF
taking into account the ring geometry
We assume that R is diagonal and don't take it into account}
\usage{
ring_partition(H, taper, block.size)
}
\arguments{
\item{H}{the observation operator}

\item{taper}{the tapering matrix (qxq)}

\item{block.size}{chunk size in domain space to take into account for one block}
}
\description{
Partition the observations for the block-LEnKPF
taking into account the ring geometry
We assume that R is diagonal and don't take it into account
}
\examples{
ndim <- 100
taper <- ring_GC(ndim, 5)
H <- diag(ndim)
partition <- ring_partition(H, taper, 5)
partition <- ring_partition( matrix( c(rep(0,ndim-1),1),1,ndim), taper, 5)
H <- matrix(0, 1, ndim)
H[1,round(ndim/2)] <- 1
partition <- ring_partition(H, taper, 5)
}
