% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics_sweq.R
\name{sweq_as_df}
\alias{sweq_as_df}
\title{from sweq_simulate to data_frame}
\usage{
sweq_as_df(state, field_names = c("fluid height", "rain content", "wind"))
}
\arguments{
\item{state}{as in sweq_simulate$state.ts}

\item{field}{names to assign}
}
\description{
take a state object from a sweq_simulate list and return a nice data frame
}
\examples{
sweq.run <- sweq_simulate(1, 60, 168)
ens0 <- sweq_ens0(40, sweq.run)
state_df <- sweq_as_df(sweq.run$state.ts[1,])
ens_df <- sweq_as_df(ens0)
#test:
all.equal(ens0[,1],  filter(ens_df, ensemble=='ens_1')$value)
}
