% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics_sweq.R
\name{sweq_change_R}
\alias{sweq_change_R}
\title{Change R matrix}
\usage{
sweq_change_R(y, R.sigr, R.sigu, ndim)
}
\arguments{
\item{y}{an observation object as retur by .sweq_allobs}

\item{R.sigr}{R.sigu are used to create the R matrix (not necessarily equal to sigr and sigu)}

\item{ndim}{dimension of the 1-dim domain}
}
\description{
take y, an observation object, and change its R matrix
(useful if needs to be changed a posteriori)
}
