% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics_sweq.R
\name{sweq_ens0}
\alias{sweq_ens0}
\title{Generate an initial ensemble for assimilation:}
\usage{
sweq_ens0(K, sweq.run, klag = 1000, umean = 0, ...)
}
\arguments{
\item{K}{Ensemble size}

\item{sweq.run}{model run list returned by sweq_simulate}

\item{klag}{number of integration steps between two members. Increase for more independence (but slower).}

\item{umean}{Initial wind field}
}
\value{
ens0
}
\description{
Generate an initial ensemble for assimilation:
}
\examples{
ndim <- 168
freq <- 60
duration <- 1*freq
sweq.run <- sweq_simulate(duration, freq, ndim)
ens0 <- sweq_ens0(K=40, sweq.run)
sweq_ggplot(sweq.run$state.ts[1,], ens0)
}
