% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{sweq_eval}
\alias{sweq_eval}
\title{Compute diagnostics of SWEQ cycling experiment}
\usage{
sweq_eval(sweq_run, model_run)
}
\arguments{
\item{sweq_run}{reference run as returned from l**_simulate}

\item{model_run}{as returned from a cycling experiment}
}
\value{
tbl object with bias, mse, rmse, nobs, crps for background/analysis ensembles
}
\description{
compute error measures (bias, mse, rmse, crps) per time and 
per ensemble type (analysis, forecast)
}
\examples{
set.seed(1)
sweq_run <- sweq_simulate(5*60, 60, 300)
k <- 20
ens0 <- sweq_ens0(k, sweq_run, klag=1000)
l <- 5
block_run <- da_cycle(ens0, sweq_run, block_LEnKPF, l=l, block_size=l/2, 
                      taper=sweq_GC(sweq_run$ndim, l, 0.9), ndim=sweq_run$ndim,
                      get_partition = sweq_partition,
                      manifold_proj = sweq_proj)
block_df <- sweq_eval(sweq_run, block_run)
}
