% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{sweq_GC}
\alias{sweq_GC}
\title{GC taper on SWEQ}
\usage{
sweq_GC(ndim, c, cross.vec = rep(0.9, 3))
}
\value{
taper matrix of size 3ndim x 3ndim
}
\description{
Compute the GC taper on the sweq geometry for h,u and r, with support half-length c
Cross correlations between fields can be defined (and should be non-zero)
}
\examples{
taper <- sweq_GC(100, 10)
image_mat(taper)
}
