% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics_sweq.R
\name{sweq_ggplot}
\alias{sweq_ggplot}
\title{Plot SWEQ}
\usage{
sweq_ggplot(state, ens = NULL, obs, h_lim = c(90.02, 90.3),
  selected_field = c(1, 2, 3), params = NULL, tit = NULL,
  norain = FALSE, h_ylim = c(89.9, 90.6), r_ylim = c(0, 0.075),
  u_ylim = c(-0.045, 0.045), field_names = c("fluid height", "rain content",
  "wind"), psize = 1)
}
\arguments{
\item{ens}{the ensemble matrix}

\item{obs}{the observation object}

\item{selected_field}{to plot only some}

\item{params}{as from sweq_simulate$params}

\item{tit}{optional title}

\item{field_names}{for plotting}

\item{psize}{size of observations}

\item{the}{true state}

\item{hlim}{the vector of hr,hc to plot as dotted lines}

\item{norain=T}{means that we plot the no-rain observations}

\item{*_ylim}{for enforcing plotting limits}
}
\description{
plot the state, the ensemble and the observations all together
}
