% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{sweq_neighbours}
\alias{sweq_neighbours}
\title{SWEQ neighbours}
\usage{
sweq_neighbours(i, l, ndim)
}
\arguments{
\item{i}{the reference position}

\item{l}{the number of neighbours in each direction}
}
\value{
a list with x.ind the indices of the neighbours and center the position i for h, u and r (in the vector state)
local.center are the indices of the centers in the local space (x.local) to apply the update
}
\description{
Find the neighbours around i, l in each direction
It takes into account the ring structure of the space and the fact that the states has three components (h,u and r)
}
\examples{
sweq_neighbours(5, 2, 10)
}
