% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{sweq_partition}
\alias{sweq_partition}
\title{Partition the observations for the block-LEnKPF
taking into account the sweq geometry
We assume that R is diagonal and don't take it into account}
\usage{
sweq_partition(H, taper, block.size)
}
\arguments{
\item{H}{the observation operator}

\item{taper}{the tapering matrix (qxq)}

\item{block.size}{chunk size in domain space to take into account for one block}
}
\description{
Partition the observations for the block-LEnKPF
taking into account the sweq geometry
We assume that R is diagonal and don't take it into account
}
\examples{
ndim <- 168
## observe all rain + some wind:
y.ind <- c( (ndim+1):(ndim+10), (2*ndim+1):(3*ndim) )
H <- diag(ndim*3)[y.ind, ]
taper <- sweq_GC(ndim, 5)
partition <- sweq_partition(H, taper, 21)
}
