% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{sweq_proj}
\alias{sweq_proj}
\title{Project back onto the manifold where rain>0:}
\usage{
sweq_proj(xa, xb, model.run, ndim = model.run$ndim)
}
\arguments{
\item{xa}{matrix of analysis ensemble to project}

\item{xb}{matrix of background ensemble (currently not used)}

\item{model.run}{object (currently not used)}

\item{ndim}{domain dimension}
}
\value{
xa with rain set to zero
}
\description{
Project back onto the manifold where rain>0:
}
