% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics_sweq.R
\name{sweq_run_as_df}
\alias{sweq_run_as_df}
\title{From sweq_run to data_frame}
\usage{
sweq_run_as_df(model_run, duration, freq, sweq_run = NULL)
}
\arguments{
\item{model_run}{returned by da_cycle}

\item{duration, freq}{information from the run}

\item{sweq_run}{true run, if passed then added with method name=state}
}
\description{
same as sweq_as_df but applied to a cycled assimilation model run
as returned by da_cycle
}
