% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{sweq_split}
\alias{sweq_split}
\title{Split a state vector into the h,u and r components
Basically the opposite of sweq_stack}
\usage{
sweq_split(state, ndim, names.only = FALSE)
}
\arguments{
\item{state}{Vector (3ndim) or matrix (3ndim x N) if ensemble}

\item{ndim}{number of dimensions in physical space}

\item{names.only}{T/F create factors (useful in  sweq_plot)}
}
\value{
a list with components h, u and r, each of size ndim
}
\description{
Split a state vector into the h,u and r components
Basically the opposite of sweq_stack
}
\examples{
state0 <- .sweq_init(10, umean=0, unoise=FALSE, topo=1)
hur <- sweq_split(state0$state, 10)
h <- hur$h
u <- hur$u
r <- hur$r
}
