% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geom.R
\name{sweq_stack}
\alias{sweq_stack}
\title{Split a state vector into the h,u and r components
Basically the oposite of sweq_split.}
\usage{
sweq_stack(h, u, r)
}
\arguments{
\item{h, u, }{r}
}
\value{
state vector or matrix
}
\description{
Split a state vector into the h,u and r components
Basically the oposite of sweq_split.
}
\examples{
state0 <- .sweq_init(10, umean=0, unoise=FALSE, topo=1)
state <- sweq_stack(state0$h, state0$u, state0$r)
}
